requirements.txt
paramiko
pyfiglet
cryptography
python-dotenv